rename: 마이페이지 관련 컴포넌트 -> 공통 컴포넌트 파일로 이동 (사용자 페이지 사용 위해) #42
