+++ app/(member)/my/page.tsx
@@ -1,6 +1,6 @@
 import UserReviewHeatmapContainer from './containers/UserReviewHeatmapContainer';
 import UserSummaryContainer from './containers/UserSummaryContainer';
-import MyActivityList from './components/MyActivityList';
+import MyActivityList from '../../components/MyActivityList';
 
 export default function Mypage() {
   return (
